+++ tests/Feature/ModelsCommandTest.php
@@ -4,6 +4,8 @@ namespace Soyhuce\NextIdeHelper\Tests\Feature;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Facades\File;
+use Soyhuce\NextIdeHelper\Tests\Fixtures\Address;
+use Soyhuce\NextIdeHelper\Tests\Fixtures\Blog\Post;
 use Soyhuce\NextIdeHelper\Tests\ResetsFixtures;
 use Soyhuce\NextIdeHelper\Tests\TestCase;
 
@@ -31,6 +33,12 @@ class ModelsCommandTest extends TestCase
             'next-ide-helper.models' => [
                 'directories' => [$this->fixturePath()],
                 'file_name' => $this->fixturePath() . '/_ide_models.php',
+                'overrides' => [
+                    Post::class => [
+                        'likes' => 'int',
+                        'address' => '?' . Address::class,
+                    ],
+                ],
             ],
         ]);